revert parser to the default for now
